fix UI + update UI for RSA + exceptions class

+++ RSA.h
@@ -11,17 +11,17 @@ class rsa {
 public:
     rsa(const int em, const std::string dm, const int im, const int ee, const int dd, const int nn) : encoded_message(em),decoded_message(dm),int_message(im),e(ee),d(dd),n(nn){}
     ~rsa(){}
-    int getEncoded(){return encoded_message;}
+    BigInt getEncoded(){return encoded_message;}
     std::string getDecoded(){return decoded_message;}
-    int get_m(){return int_message;}
-    int get_e(){return e;}
-    int get_d(){return d;}
-    int get_n(){return n;}
-    void setEncoded(const int em){encoded_message=em;}
+    BigInt get_m(){return int_message;}
+    BigInt get_e(){return e;}
+    BigInt get_d(){return d;}
+    BigInt get_n(){return n;}
+    void setEncoded(const BigInt em){encoded_message=em;}
     void setDecoded(const std::string dm){decoded_message=dm;}
-    void set_m(const int im){int_message=im;}
-    void set_e(const int ee){e=ee;}
-    void set_d(const int dd){d=dd;}
+    void set_m(const BigInt im){int_message=im;}
+    void set_e(const BigInt ee){e=ee;}
+    void set_d(const BigInt dd){d=dd;}
     void set_n(const int nn){n=nn;}
     int calc_n(const int p, const int q){set_n(p*q);return p*q;}
     int calc_phi_n(const int p, const int q);
@@ -32,11 +32,11 @@ public:
     BigInt M_to_m(const std::string message); //OS2IP from RFC
     std::string m_to_M(const BigInt message); //OS2IP reverse operation
 private:
-    int encoded_message;
+    BigInt encoded_message;
     std::string decoded_message;
-    int int_message;
-    int e;
-    int d;
+    BigInt int_message;
+    BigInt e;
+    BigInt d;
     int n;
 };
